update style for components shared/candidacy/*
