+++ resources/js/shared/candidacy/candidacy-card.tsx
